cityzen_dashboard-main/requirements.txt
contractions==0.1.66
docx2txt==0.8
emoji==1.6.3
gensim==4.1.2
nltk==3.7
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.1
pdfminer.six==20191110
pdfreader==0.1.10
plotly==5.7.0
protobuf==3.20.1
pymupdf==1.19.6
pypandoc==1.7.2
pypdf2==1.26.0
python-docx==0.8.11
python-pptx==0.6.21
#streamlit==1.6.0
streamlit==1.8.1
syllables==1.0.3
torch==1.11.0
unidecode==1.3.3
